+++ packages/plugin-package-commonjs/src/index.ts
@@ -1,1 +1,0 @@
-export {buildCommonJsOutput} from './plugin-package-commonjs';