+++ packages/plugin-package-esmodules/src/index.ts
@@ -1,4 +1,0 @@
-export {
-  esmodulesOutput,
-  buildEsModulesOutput,
-} from './plugin-package-esmodules';