+++ packages/plugin-package-node/src/index.ts
@@ -1,1 +1,0 @@
-export {buildNodeOutput, nodeOutput} from './plugin-package-node';