+++ packages/plugin-package-typescript/src/index.ts
@@ -1,1 +1,0 @@
-export {buildTypeScriptDefinitions} from './plugin-package-typescript';